feat: organize wrap behavior
