+++ src/judgeval/common/tracer/__init__.py
@@ -2,17 +2,19 @@ from judgeval.common.tracer.core import (
     TraceClient,
     _DeepTracer,
     Tracer,
-    wrap,
     current_span_var,
     current_trace_var,
     SpanType,
-    cost_per_token,
 )
 from judgeval.common.tracer.otel_exporter import JudgmentAPISpanExporter
 from judgeval.common.tracer.otel_span_processor import JudgmentSpanProcessor
 from judgeval.common.tracer.span_processor import SpanProcessorBase
 from judgeval.common.tracer.trace_manager import TraceManagerClient
 from judgeval.data import TraceSpan
+from judgeval.common.tracer.llm_wrapper import (
+    wrap,
+    cost_per_token,
+)
 
 __all__ = [
     "_DeepTracer",
